Improve README.md

+++ asn1-utils/src/simple.rs
@@ -48,7 +48,7 @@ ber_newtype!(SomeString);
 #[derive(Debug)]
 struct IntSequence {
   a: u64,
-  b: Vec<u64>,
+  b: Vec<i32>,
   c: SomeString,
 }
